requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==39.0.1
defusedxml==0.7.1
distlib==0.3.6
Django==3.2.5
django-allauth==0.45.0
django-cors-headers==3.14.0
django-crispy-forms==1.8.1
djangorestframework==3.14.0
filelock==3.9.0
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
Pillow==9.3.0
platformdirs==2.6.2
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.26.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
simplejson==3.18.0
sqlparse==0.4.3
urllib3==1.24.2
virtualenv==20.16.7
Werkzeug==2.1.2
whitenoise==6.2.0
